feat: Add additional step to render ytt resources
